requirements.txt
asgiref==3.8.1
Django==5.0.14
django-filter==25.1
django-mssql-backend==2.8.1
djangorestframework==3.16.0
mssql-django==1.5
numpy==2.2.6
pandas==2.2.3
pyodbc==5.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2